HHH-7318 Fix native queries incorrectly converting BIGINT results to BigInteger
Also fix existing tests with wrong expectations:

* TupleNativeQueryTest: there is no reason to expect a BigInteger in
results, since the ID of entity Users if of type Long.
* MultiLineImportFileTest: there is no reason to expect "COUNT(*)" to
return a BigInteger, since the SQL type of COUNT(*) is bigint, which
corresponds to Long in Java.

+++ hibernate-core/src/test/java/org/hibernate/test/fileimport/MultiLineImportFileTest.java
@@ -56,7 +56,7 @@ public class MultiLineImportFileTest extends BaseCoreFunctionalTestCase {
 		Session s = openSession();
 		final Transaction tx = s.beginTransaction();
 
-		BigInteger count = (BigInteger) s.createNativeQuery( "SELECT COUNT(*) FROM test_data" ).uniqueResult();
+		Long count = (Long) s.createNativeQuery( "SELECT COUNT(*) FROM test_data" ).uniqueResult();
 		assertEquals( "Incorrect row number", 3L, count.longValue() );
 
 		final String multiLineText = (String) s.createNativeQuery( "SELECT text FROM test_data WHERE id = 2" )
